allow for launching without reload
