+++ packages/plugin-core/src/commands/LaunchTutorialWorkspaceCommand.ts
@@ -45,6 +45,7 @@ export class LaunchTutorialWorkspaceCommand extends BasicCommand<
       rootDirRaw: filePath,
       workspaceInitializer: new TutorialInitializer(),
       workspaceType: WorkspaceType.NATIVE,
+      EXPERIMENTAL_openNativeWorkspaceNoReload: true,
     });
   }
 }
